// src/models/ICallback.tsx
export interface ICallBack {
    full_name: string
    phone_number: string
    start_time: string
    end_time: string
    comment: string
}
export interface IResponseCallBack {
    full_name: string
    phone_number: string
    start_time: string
    end_time: string
    comment: string
    id: number
}